Rename mod to "Investiture"
This reflects the nature of this project way better, the final goal being a complete realisation of Sanderson's Cosmere in Minecraft. In code, this change is represented by a modularisation of the project, the first module being "Allomancy".
