--- src/main/java/de/mineformers/allomancy/client/KeyBindings.java
@@ -1,15 +1,0 @@
-package de.mineformers.allomancy.client;
-
-import net.minecraft.client.settings.KeyBinding;
-import org.lwjgl.input.Keyboard;
-
-/**
- * KeyBindings
- *
- * @author PaleoCrafter
- */
-public class KeyBindings
-{
-    public static final KeyBinding SHOW_DIAL =
-        new KeyBinding("key.showDial", Keyboard.KEY_F, "key.categories.allomancy");
-}
