--- src/main/java/de/mineformers/allomancy/core/ServerProxy.java
@@ -1,10 +1,0 @@
-package de.mineformers.allomancy.core;
-
-/**
- * ServerProxy
- *
- * @author PaleoCrafter
- */
-public class ServerProxy implements Proxy
-{
-}
