--- src/main/java/de/mineformers/allomancy/metal/MetalEffects.java
@@ -1,13 +1,0 @@
-package de.mineformers.allomancy.metal;
-
-/**
- * MetalEffects
- *
- * @author PaleoCrafter
- */
-public interface MetalEffects extends AllomanticMetal
-{
-    void startBurning(MetalBurner burner);
-
-    void stopBurning(MetalBurner burner);
-}
