--- src/main/java/de/mineformers/allomancy/network/messages/ToggleBurningMetal.java
@@ -1,22 +1,0 @@
-package de.mineformers.allomancy.network.messages;
-
-import de.mineformers.allomancy.network.Message;
-
-/**
- * ToggleBurningMetal
- *
- * @author PaleoCrafter
- */
-public class ToggleBurningMetal extends Message
-{
-    public String metal;
-
-    public ToggleBurningMetal()
-    {
-    }
-
-    public ToggleBurningMetal(String metal)
-    {
-        this.metal = metal;
-    }
-}
