+++ src/main/java/de/mineformers/investiture/allomancy/core/ServerProxy.java
@@ -1,0 +1,12 @@
+package de.mineformers.investiture.allomancy.core;
+
+import de.mineformers.investiture.core.Proxy;
+
+/**
+ * ServerProxy
+ *
+ * @author PaleoCrafter
+ */
+public class ServerProxy implements Proxy
+{
+}
