+++ src/main/java/de/mineformers/investiture/allomancy/metal/MetalEffects.java
@@ -1,0 +1,13 @@
+package de.mineformers.investiture.allomancy.metal;
+
+/**
+ * MetalEffects
+ *
+ * @author PaleoCrafter
+ */
+public interface MetalEffects extends AllomanticMetal
+{
+    void startBurning(MetalBurner burner);
+
+    void stopBurning(MetalBurner burner);
+}
