+++ src/main/java/de/mineformers/investiture/client/KeyBindings.java
@@ -1,0 +1,15 @@
+package de.mineformers.investiture.client;
+
+import net.minecraft.client.settings.KeyBinding;
+import org.lwjgl.input.Keyboard;
+
+/**
+ * KeyBindings
+ *
+ * @author PaleoCrafter
+ */
+public class KeyBindings
+{
+    public static final KeyBinding SHOW_DIAL =
+        new KeyBinding("key.showDial", Keyboard.KEY_F, "key.categories.allomancy");
+}
