+++ src/main/java/de/mineformers/investiture/core/ClientProxy.java
@@ -1,0 +1,10 @@
+package de.mineformers.investiture.core;
+
+/**
+ * ClientProxy
+ *
+ * @author PaleoCrafter
+ */
+public class ClientProxy implements Proxy
+{
+}
