+++ src/main/java/de/mineformers/investiture/core/Manifestation.java
@@ -1,0 +1,19 @@
+package de.mineformers.investiture.core;
+
+import net.minecraftforge.fml.common.event.FMLInitializationEvent;
+import net.minecraftforge.fml.common.event.FMLPostInitializationEvent;
+import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
+
+/**
+ * Manifestation
+ *
+ * @author PaleoCrafter
+ */
+public interface Manifestation
+{
+    void preInit(FMLPreInitializationEvent event);
+
+    void init(FMLInitializationEvent event);
+
+    void postInit(FMLPostInitializationEvent event);
+}
