+++ src/main/java/de/mineformers/investiture/core/Proxy.java
@@ -1,0 +1,25 @@
+package de.mineformers.investiture.core;
+
+import net.minecraftforge.fml.common.event.FMLInitializationEvent;
+import net.minecraftforge.fml.common.event.FMLPostInitializationEvent;
+import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
+
+/**
+ * Proxy
+ *
+ * @author PaleoCrafter
+ */
+public interface Proxy
+{
+    default void preInit(FMLPreInitializationEvent event)
+    {
+    }
+
+    default void init(FMLInitializationEvent event)
+    {
+    }
+
+    default void postInit(FMLPostInitializationEvent event)
+    {
+    }
+}
