+++ src/main/java/de/mineformers/investiture/core/ServerProxy.java
@@ -1,0 +1,10 @@
+package de.mineformers.investiture.core;
+
+/**
+ * ServerProxy
+ *
+ * @author PaleoCrafter
+ */
+public class ServerProxy implements Proxy
+{
+}
