+++ src/main/resources/assets/allomancy/lang/en_US.lang
@@ -1,4 +1,3 @@
-itemGroup.allomancy=Allomancy
 allomancy.message.purity=%s%% pure
 
 allomancy.metals.bronze.name=Bronze
