+++ src/main/resources/assets/investiture/lang/en_US.lang
@@ -1,0 +1,1 @@
+itemGroup.investiture=Investiture
